verificação de erros php

+++ cadastro.php
@@ -9,6 +9,7 @@ if (!isset($_GET['usuario'])) {
             </script>
         ';
 }
+$cpferro = FALSE;
 ?>
 <html lang="en">
 
@@ -44,14 +45,47 @@ if (!isset($_GET['usuario'])) {
     <script src="vendor/select2/select2.min.js"></script>
     <!--===============================================================================================-->
     <script src="js/main.js"></script>
-    <script src="js/jquery.maskedinput.js"></script>
+    <!-- <script src="js/jquery.maskedinput.js"></script> -->
 </head>
 
 <body>
     <?php
-    if (empty($_POST['CPF'])) {
-        $cpferro = "* campo necessário";
-        $existe_erro = TRUE;
+    if ($_SERVER["REQUEST_METHOD"] == "POST") {
+        if (empty($_POST['nome'])) {
+            $nomeerro = '
+            <span style="color: #ff0000">
+            * campo obrigatório
+            </span>';
+            $existe_erro = TRUE;
+        }
+        if (empty($_POST['CPF'])) {
+            $cpferro = '
+            <span style="color: #ff0000">
+            * campo obrigatório
+            </span>';
+            $existe_erro = TRUE;
+        }
+        if (empty($_POST['email'])) {
+            $emailerro = '
+            <span style="color: #ff0000">
+            * campo obrigatório
+            </span>';
+            $existe_erro = TRUE;
+        }
+        if (empty($_POST['senha'])) {
+            $senhaerro = '
+            <span style="color: #ff0000">
+            * campo obrigatório
+            </span>';
+            $existe_erro = TRUE;
+        }
+        if (!(isset($_POST['termos']))) {
+            $termoserro = '
+            <span style="color: #ff0000">
+            * necessário concordar com os termos de uso.
+            </span>';
+            $existe_erro = TRUE;
+        }
     }
     ?>
     <div class="limiter">
@@ -77,27 +111,34 @@ if (!isset($_GET['usuario'])) {
                     </div>
 
                     <div class="wrap-input100 m-b-16">
-                        <input class="input100" type="text" name="nome" placeholder="Nome" id="nome">
+                        <input class="input100" type="text" name="nome" placeholder="Nome" id="nome" value="<?php echo isset($_POST['nome']) ? $_POST['nome'] : '' ?>">
                         <span class="focus-input100"></span>
                         <span class="symbol-input100"></span>
+                        <?php
+                        if (isset($nomeerro)) echo $nomeerro;
+                        ?>
                     </div>
                     <div class="wrap-input100 m-b-16">
-                        <input class="input100" type="text" name="CPF" placeholder="CPF" id="cpf">
+                        <input class="input100" type="text" name="CPF" placeholder="CPF" id="cpf" value="<?php echo isset($_POST['CPF']) ? $_POST['CPF'] : '' ?>">
                         <span class="focus-input100"></span>
                         <span clfass="symbol-input100"></span>
-                        <span>
-                            <?php
-                            if (isset($cpferro)) echo $cpferro;
-                            ?>
-                        </span>
+                        <?php
+                        if (isset($cpferro)) echo $cpferro;
+                        ?>
                     </div>
                     <div class="wrap-input100 m-b-16">
-                        <input class="input100" type="email" name="email" placeholder="Email" id="email">
+                        <input class="input100" type="email" name="email" placeholder="Email" id="email" value="<?php echo isset($_POST['email']) ? $_POST['email'] : '' ?>">
                         <span class="focus-input100"></span>
                         <span class="symbol-input100"></span>
+                        <?php
+                        if (isset($emailerro)) echo $emailerro;
+                        ?>
                     </div>
                     <span class="wrap-input100 validate-input m-b-16">
                         <input class="input100" type="password" name="senha" placeholder="Senha" id="senha">
+                        <?php
+                        if (isset($senhaerro)) echo $senhaerro;
+                        ?>
                     </span>
 
                     <?php
@@ -121,10 +162,13 @@ if (!isset($_GET['usuario'])) {
                     ?>
 
                     <div class="contact100-form-checkbox m-l-4">
-                        <input class="input-checkbox100" id="ckb1" type="checkbox" name="remember-me">
+                        <input class="input-checkbox100" id="ckb1" type="checkbox" name="termos">
                         <label class="label-checkbox100" for="ckb1">
                             <span class="txt1">Eu concordo com os </span><a class="txt1 bo1 hov1" href="termosdeservico.html">Termos de serviço</a>
                         </label>
+                        <?php
+                        if (isset($termoserro)) echo $termoserro;
+                        ?>
                     </div>
 
                     <div class="container-login100-form-btn p-t-10">
@@ -146,51 +190,53 @@ if (!isset($_GET['usuario'])) {
         </div>
     </div>
     <?php
-    if (isset($_POST['insert'])) {
-        $nome     = $_POST['nome'];
-        $cpf       = $_POST['CPF'];
-        $email     = $_POST['email'];
-        $senha     = $_POST['senha'];
-        if ($usu == 0) {
-            $sql = "insert into USUARIOS (user_cpf,user_nome,user_email,user_senha, user_tipo)
+        if (@$existe_erro != TRUE) {
+            if (isset($_POST['insert'])) {
+                $nome     = $_POST['nome'];
+                $cpf       = $_POST['CPF'];
+                $email     = $_POST['email'];
+                $senha     = $_POST['senha'];
+                if ($usu == 0) {
+                    $sql = "insert into USUARIOS (user_cpf,user_nome,user_email,user_senha, user_tipo)
 	    			VALUES
                 ('$cpf','$nome','$email','$senha',$usu)";
-            $executar = sqlsrv_query($con, $sql);
-        }
-        if ($usu == 1) {
-            $RG = $_POST['rg'];
-            $orgemissor = $_POST['orgao'];
-            $nomemae = $_POST['nomemae'];
-            $nomepai = $_POST['nomepai'];
-            $sql = "insert into USUARIOS (user_cpf,user_nome,user_email,user_senha, user_tipo)
+                    $executar = sqlsrv_query($con, $sql);
+                }
+                if ($usu == 1) {
+                    $RG = $_POST['rg'];
+                    $orgemissor = $_POST['orgao'];
+                    $nomemae = $_POST['nomemae'];
+                    $nomepai = $_POST['nomepai'];
+                    $sql = "insert into USUARIOS (user_cpf,user_nome,user_email,user_senha, user_tipo)
                 VALUES
                 ('$cpf','$nome','$email','$senha',$usu)";
-            $executar = sqlsrv_query($con, $sql);
-            // pegar id do entregador
-            $entregadorID = sqlsrv_query($con, "SELECT user_codigo FROM USUARIOS WHERE user_cpf = '$cpf'");
-            $entregadorID2 = implode("", sqlsrv_fetch_array($entregadorID, SQLSRV_FETCH_ASSOC));
-            echo $entregadorID2;
-            // echo $entregadorID;
-            // $entregadorID2 = $entregadorID2array['user_codigo'];
-            $sqlentregador = "insert into ENTREGADORES (user_codigo,entregador_rg,entregador_orgao_emissor,entregador_nomemae,entregador_nomepai)
+                    $executar = sqlsrv_query($con, $sql);
+                    // pegar id do entregador
+                    $entregadorID = sqlsrv_query($con, "SELECT user_codigo FROM USUARIOS WHERE user_cpf = '$cpf'");
+                    $entregadorID2 = implode("", sqlsrv_fetch_array($entregadorID, SQLSRV_FETCH_ASSOC));
+                    echo $entregadorID2;
+                    // echo $entregadorID;
+                    // $entregadorID2 = $entregadorID2array['user_codigo'];
+                    $sqlentregador = "insert into ENTREGADORES (user_codigo,entregador_rg,entregador_orgao_emissor,entregador_nomemae,entregador_nomepai)
                 VALUES
                 ($entregadorID2,'$RG','$orgemissor','$nomemae','$nomepai')";
-            // echo $sqlentregador;
-            $executarentregador = sqlsrv_query($con, $sqlentregador);
-        }
+                    // echo $sqlentregador;
+                    $executarentregador = sqlsrv_query($con, $sqlentregador);
+                }
 
 
-        if ($executar) {
-            echo "<script>alert('Usuário inserido corretamente')</script>";
-        } else {
-            echo "nao inserido";
-        }
-        if (isset($executarentregador)) {
-            if ($executarentregador) {
-                echo "<script>alert('Entregador inserido corretamente')</script>";
+                if ($executar) {
+                    echo "<script>alert('Usuário inserido corretamente')</script>";
+                } else {
+                    echo "nao inserido";
+                }
+                if (isset($executarentregador)) {
+                    if ($executarentregador) {
+                        echo "<script>alert('Entregador inserido corretamente')</script>";
+                    }
+                }
             }
         }
-    }
     ?>
 </body>
